requirements.txt
easydict==1.9
kornia==0.3.1
tensorboardX==2.0
tqdm==4.41.1
easydict==1.7
graphviz==0.8.4
imageio==2.3.0
matplotlib==2.2.2
numpy==1.14.5
Pillow==5.2.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
tensorboardX==1.2
torch==1.4.0
torchvision==0.5.0
